rename: set dir
