+++ project_today/lib/core/constant/index.dart
@@ -1,0 +1,2 @@
+export 'colors.dart';
+export 'fonts.dart';
