+++ project_today/lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:project_today/screen/screens_group.dart';
-import 'package:project_today/screen/screens_diary.dart';
-import 'package:project_today/screen/screens_onboard.dart';
+import 'package:project_today/screen/group/group_screens.dart';
+import 'package:project_today/screen/diary/diary_screens.dart';
+import 'package:project_today/screen/onboard/onboard_screens.dart';
 import 'package:project_today/screen/screens_index.dart';
-import 'package:project_today/screen/screens_splash.dart';
+import 'package:project_today/screen/splash/splash_screens.dart';
 
 void main() {
   runApp(const MyApp());
